+++ src/components/posts/post-section.tsx
@@ -14,7 +14,9 @@ export function PostSection({ title, href, headerChildren, children }: IProps) {
   return (
     <section className="flex flex-col gap-y-8">
       <VCenterRow className="border-t-3 border-b border-foreground/75 py-3 justify-between">
-        <h2>{href ? <BaseLink href={href}>{title}</BaseLink> : title}</h2>
+        <h2 className="font-bold">
+          {href ? <BaseLink href={href}>{title}</BaseLink> : title}
+        </h2>
         <VCenterRow>{headerChildren && headerChildren}</VCenterRow>
       </VCenterRow>