+++ src/components/posts/post-title-link.tsx
@@ -6,7 +6,7 @@ import { getPostUrl } from '@/lib/post'
 import type { IPostProps } from './post-tags'
 
 const CLS = cn(
-  'group-hover:underline decoration-2 underline-offset-4 group-hover:text-sky-700 group-hover:decoration-sky-700',
+  'group-hover:underline decoration-2 underline-offset-4 group-hover:text-sky-600 group-hover:decoration-sky-600',
   'data-[mode=dark]:text-white data-[mode=dark]:group-hover:text-sky-400 data-[mode=dark]:group-hover:decoration-sky-400'
 )