+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
+- Fixed index and people pages and expanded footer.
 - Added pinned feature posts and adjust layout of pages.
 - UI fixes.
 - UI fixes and standardization.